ML/source/requirements.txt
torch
transformers
pydub
librosa
fpdf
python-docx
numpy
requests
fastapi
pydantic
uvicorn
